Еще исправления замечаний
