+++ project/src/components/app/app.tsx
@@ -1,19 +1,13 @@
 import MainPage from '../../pages/main/main';
-type PromoFilmProps = {
-  title: string,
-  genre: string,
-  date: number,
-  src: string
-}
-
+import { PromoFilmProps } from '../../types/types';
 
 function App({title, genre, date, src}: PromoFilmProps): JSX.Element {
   return (
     <MainPage
-      title = {title}
-      genre = {genre}
-      date = {date}
-      src = {src}
+      title={title}
+      genre={genre}
+      date={date}
+      src={src}
     />
   );
 }
