+++ project/src/data.ts
@@ -1,3 +1,3 @@
-const RATING: number[] = [1,2,3,4,5,6,7,8,9,10];
+const RATING_VALUES: number[] = [1,2,3,4,5,6,7,8,9,10];
 
-export {RATING};
+export {RATING_VALUES};
