+++ project/src/types/types.ts
@@ -1,0 +1,11 @@
+export type Film = {
+  src: string,
+  title: string
+}
+
+export type PromoFilmProps = {
+  title: string,
+  genre: string,
+  date: number,
+  src: string
+}
